Relicense under GPLv3
Previously, it is indicated that the schema "on the whole" is licensed
under CC-BY but with per-file exceptions. This is confusing for
several reasons.

1. It is easy to be misinterpreted (maybe on purpose) that everything
in the project is licensed under CC-BY, which is definitely not the
case.

2. The packaged distribution itself cannot be licensed under CC-BY
because of the stricter GPL-licensed components, which are required
for rime-moran.

This relicensing does not override the licensing decisions of many Lua
components and dictionaries. Rather, it only tries to indicate that
the distribution itself is GPL-licensed. At the same time, the chai
file is relicensed under a stricter license, because it's the right
thing to do. Other files without copyright notice must be understood
as "provided by rime-moran" and should be considered
GPL-licensed. Please contact us if you would need a more relaxed
license.

And, hey, not everyone obeys even the super relaxed CC-BY. Why not
just use a stricter one? ;-)